+++ src/heuristiques/Heuristica.java
@@ -4,6 +4,6 @@ import model.Casella;
 
 public abstract class Heuristica {
 
-    public abstract float calcular(Casella actual, Casella objectiu);
+    public abstract double calcular(Casella actual, Casella objectiu);
     
 }
