+++ src/heuristiques/HeuristicaCanvi.java
@@ -5,9 +5,9 @@ import model.Casella;
 public class HeuristicaCanvi extends Heuristica {
     
     @Override
-    public float calcular(Casella actual, Casella objectiu) {
+    public double calcular(Casella actual, Casella objectiu) {
 
-        float h = Math.abs(actual.getX() - objectiu.getX()) +
+        double h = Math.abs(actual.getX() - objectiu.getX()) +
                   Math.abs(actual.getY() - objectiu.getY());
 
         if actual.getTipus == objectiu.getTipus {
