+++ src/heuristiques/HeuristicaTipus.java
@@ -5,7 +5,7 @@ import model.Casella;
 public class HeuristicaTipus extends Heuristica {
  
     @Override
-    public float calcular(Casella actual, Casella objectiu) {
+    public double calcular(Casella actual, Casella objectiu) {
         return [Math.abs(actual.getX() - objectiu.getX()) +
                Math.abs(actual.getY() - objectiu.getY())] *
                objectiu.getTipus();
